Add initial MIPS port, sans automatic partitioner
Help with the automatic partitioner by parties interested in the MIPS
ports of Adélie (and/or Alpine and/or postmarketOS) would be very
desirable.

I'm only currently aware myself of how to partition SGI MIPS64 machines.
My gut tells me we probably need the `subarch` system like for PPC64,
but I am not sure how non-SGI MIPS64 machines need to be partitioned.